fix: added comments and , cleand up the code

+++ delete_log.sh
@@ -1,12 +1,19 @@
 #!/bin/bash
-find /var/log -type f -mtime +2 > filelist.txt
-cat filelist.txt
+LOG_DIR="/var/log"
+FILE_LIST="filelist.txt"
+BACKUP_FILE="bk-$(date +%F).tar.gz"
+
+find ${LOG_DIR} -type f -mtime +2 > ${FILE_LIST}
+if [[ ! -s ${FILE_LIST} ]]; then
+        echo "No files to delete."
+        exit 0
+fi
 #take backup
 echo "Taking backup of the files..."
-tar -zcf "bk-$(date +%F).tar.gz"  -T filelist.txt
+tar -zcf ${BACKUP_FILE}  -T ${FILE_LIST}
 echo "Deleteing file.."
 while read -r filename;do
         rm ${filename}
-done < filelist.txt
-rm ${filelist.txt}
+done < ${FILE_LIST}
+rm ${FILE_LIST}
